Refactor config tests
Move tests of each source file into its own spec, and ensure each module
is tested at its own level.

Issue: BB-808
